test connect serializer
